MovingSensitivity
Speaking is the essential art in gathering new perspectives of life, and by new it is meant to be a progress from our ancestors who had to survive just so we could have a chance of a life to our descendants, who'll live in an uncertain future. How can you make bets on an uncertain future whom you'll never know how it is gonna be unless you have control over yourself?